move functions that we use in utils into utilities; add docs
